requirements.txt
accelerate==0.18.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
anaconda-client==1.7.2
anaconda-navigator==2.0.3
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
async-timeout==4.0.3
atomicwrites==1.4.0
attrdict==2.0.1
Babel==2.14.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bce-python-sdk==0.9.5
bkcharts==0.2
black==19.10b0
blinker==1.7.0
blis==0.7.11
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2020.12.5
chainer==7.8.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
clyent==1.2.2
cmake==3.28.1
coloredlogs==15.0.1
compel==2.0.2
conda==4.10.1
conda-build==3.21.4
conda-verify==3.4.2
confection==0.1.4
contextlib2==0.6.0.post1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.10.0
cymem==2.0.8
Cython==3.0.7
cytoolz==0.11.0
diffusers==0.23.0
docker-pycreds==0.4.0
easydict==1.11
efficientnet-pytorch==0.7.1
einops==0.7.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.0
face-alignment==1.3.0
fastapi==0.104.1
fastcache==1.1.0
ffmpy==0.3.1
fire==0.6.0
Flask==2.2.5
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.50.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.2
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.0.8
gradio==4.7.1
gradio_client==0.7.0
h11==0.14.0
h5py==2.10.0
HeapDict==1.0.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
imgaug==0.4.0
importlib-resources==6.1.1
inflect==7.0.0
insightface==0.7.3
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.3
json5==0.9.5
jupyter==1.0.0
kornia==0.7.0
langcodes==3.3.0
lightning-utilities==0.10.0
lit==15.0.7
llvmlite==0.36.0
lmdb==1.4.1
locket==0.2.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.6.1
mdurl==0.1.2
mediapy==1.2.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
multidict==6.0.4
multipledispatch==0.6.0
munch==4.0.0
murmurhash==1.0.10
mypy-extensions==0.4.3
navigator-updater==0.2.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.1
olefile==0.46
omegaconf==2.3.0
onnx==1.13.0
onnx-tool==0.8.1
onnxruntime-gpu==1.16.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.10
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle==2.5.1
pandas==1.5.3
parso==0.7.0
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pdf2docx==0.5.8
pep8==1.7.1
pillow==10.2.0
pkginfo==1.7.0
ply==3.11
premailer==3.10.0
preshed==3.0.9
pretrainedmodels==0.7.4
prettytable==3.9.0
protobuf==3.20.3
pyarrow==14.0.1
pyclipper==1.3.0.post5
pycosat==0.6.3
pycryptodome==3.20.0
pycurl==7.43.0.6
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.20.2
pyodbc===4.0.0-unsupported
pyparsing==3.1.1
pytest==6.2.3
python-docx==1.1.0
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytorch-triton==2.1.0+9820899b38
pytz==2024.1
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
qudida==0.0.4
rapidfuzz==3.7.0
rarfile==4.1
requests==2.31.0
rich==13.7.0
safetensors==0.4.0
scikit-image==0.18.1
scikit-learn==1.0
segmentation-models-pytorch==0.3.3
semantic-version==2.10.0
sentry-sdk==1.22.2
setproctitle==1.3.2
shapely==2.0.3
shellingham==1.5.4
simplegeneric==0.8.1
sip==4.19.13
smart-open==6.4.0
smmap==5.0.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
streamlit==1.29.0
tables==3.6.1
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.9.4
thinc==8.2.2
tifffile==2020.10.1
timm==0.9.2
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1
tr==2.3.1
transformers==4.27.4
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
unicodecsv==0.14.1
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
visualdl==2.5.3
wandb==0.15.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xformers==0.0.22.post7
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.9.3
zict==2.0.0
zope.event==4.5.0
